pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "asp"
version = "0.1.0"
description = "Agentic Services Protocol (ASP) — schema definitions, generation pipeline, and documentation"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "ASP Authors" },
]
dependencies = [
    "jsonschema>=4.20.0",
    "pyyaml>=6.0",
    "referencing>=0.31.0",
]

[project.optional-dependencies]
dev = [
    "datamodel-code-generator>=0.25.0",
    "ruff>=0.4.0",
    "pre-commit>=3.7.0",
]
docs = [
    "mkdocs-material>=9.5.0",
    "mkdocs-macros-plugin>=1.0.0",
]

[tool.setuptools.packages.find]
include = ["asp*", "source*", "generated*", "spec*"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["F", "W", "I"]

[tool.ruff.format]
quote-style = "double"
